Refactored the vector class.
 - Added a normalization function to normalize vectors.
 - Changed the point type to doubles to allow for accurate normalization.
 - Separated vector.hpp into two files: vector.h and vector.cpp since the normalization function couldn't be in the header.
 - Updated all files which include vector.hpp, they now include vector.h.
 - Changed a cast in src/GameStates/SinglePlayerGameState.cpp to a double to avoid the narrowing conversion warning.

+++ src/GameStates/SinglePlayerGameState.cpp
@@ -15,7 +15,10 @@ namespace pong
     this->bottomPaddle_.controller(std::shared_ptr<PaddleController>(
                                                     new MousePaddleController));
     this->bottomPaddle_.position(
-                    {0,static_cast<int>(Game::getInstance()->height()
+//                                    ,--This is casted to a double here to
+//                                   /   avoid narrowing conversion warnings. I
+//                                  \/   don't really like it but, eh.
+                    {0,static_cast<double>(Game::getInstance()->height()
                                         - this->bottomPaddle_.height())});
 
     //Configure the mouse.
